PlayerDamage
Handles taking damage, hitsounds and knockback
